+++ mono/metadata/debug-mono-symfile.h
@@ -12,6 +12,7 @@
 #include <mono/metadata/class.h>
 #include <mono/metadata/reflection.h>
 #include <mono/metadata/mono-debug.h>
+#include <mono/metadata/debug-internals.h>
 
 typedef struct MonoSymbolFileOffsetTable	MonoSymbolFileOffsetTable;
 typedef struct MonoSymbolFileLineNumberEntry	MonoSymbolFileLineNumberEntry;
@@ -71,69 +72,6 @@ struct MonoSymbolFileMethodAddress {
 	uint8_t data [MONO_ZERO_LEN_ARRAY];
 };
 
-struct _MonoDebugMethodInfo {
-	MonoMethod *method;
-	MonoDebugHandle *handle;
-	uint32_t index;
-	uint32_t data_offset;
-	uint32_t lnt_offset;
-};
-
-typedef struct {
-	int parent;
-	int type;
-	/* IL offsets */
-	int start_offset, end_offset;
-} MonoDebugCodeBlock;
-
-typedef struct {
-	char *name;
-	int index;
-	/* Might be null for the main scope */
-	MonoDebugCodeBlock *block;
-} MonoDebugLocalVar;
-
-/*
- * Information about local variables retrieved from a symbol file.
- */
-struct _MonoDebugLocalsInfo {
-	int num_locals;
-	MonoDebugLocalVar *locals;
-	int num_blocks;
-	MonoDebugCodeBlock *code_blocks;
-};
-
-/*
-* Information about method await yield and resume offsets retrieved from a symbol file.
-*/
-struct _MonoDebugMethodAsyncInfo {
-	uint32_t catch_handler_offset;
-	int num_awaits;
-	uint32_t *yield_offsets;
-	uint32_t *resume_offsets;
-	uint32_t *move_next_method_token;
-};
-
-struct _MonoDebugLineNumberEntry {
-	uint32_t il_offset;
-	uint32_t native_offset;
-};
-
-/*
- * Information about a source file retrieved from a symbol file.
- */
-typedef struct {
-	char *source_file;
-	/* 16 byte long */
-	guint8 *guid, *hash;
-} MonoDebugSourceInfo;
-
-typedef struct {
-	int il_offset;
-	int line, column;
-	int end_line, end_column;
-} MonoSymSeqPoint;
-
 #define MONO_SYMBOL_FILE_MAJOR_VERSION		50
 #define MONO_SYMBOL_FILE_MINOR_VERSION		0
 #define MONO_SYMBOL_FILE_MAGIC			0x45e82623fd7fa614ULL
@@ -168,9 +106,6 @@ mono_debug_symfile_lookup_locals (MonoDebugMethodInfo *minfo);
 
 void
 mono_debug_symfile_get_seq_points (MonoDebugMethodInfo *minfo, char **source_file, GPtrArray **source_file_list, int **source_files, MonoSymSeqPoint **seq_points, int *n_seq_points);
-
-gboolean
-mono_debug_image_has_debug_info (MonoImage *image);
 
 MONO_END_DECLS
