+++ mono/profiler/mono-profiler-vtune.c
@@ -31,7 +31,7 @@
 #include <mono/metadata/debug-helpers.h>
 #include <mono/metadata/assembly.h>
 #include <mono/metadata/mono-debug.h>
-#include <mono/metadata/debug-mono-symfile.h>
+#include <mono/metadata/debug-internals.h>
 #include <string.h>
 #include <glib.h>
